chore: Added sequenceS and modularized the confirm screen
